Serial output select menu entry / Serial selection to variant
Added a menu entry to the Arduino IDE to select between the Serial
interface to the Host PC and on board. Preprocessor defines
SERIAL_HOSTPC/SERIAL_ONBOARD are introduced and conditionally
checked in pins_arduino.h for the following kits:

- XMC1100 Boot Kit
- XMC1100 XMC2Go
- XMC1100 H-Bridge 2Go
- XMC1300 Boot Kit

Default selection of none of these are defined is to set SERIAL_DEBUG to 1.

Added selection of serial interface via #define to XMC1100 H-Bridge 2Go.